requirements.txt
Flask==2.3.0
SQLAlchemy==1.3.2
pymysql==1.1.0
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.10.4